% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min-hash-for-file.R
\name{as.character.lzjd_hash}
\alias{as.character.lzjd_hash}
\title{Turn an \code{lzhd_hash} object into a base 64 encoded string}
\usage{
\method{as.character}{lzjd_hash}(x, ...)
}
\arguments{
\item{x, }{an \code{lzhd_hash} object}

\item{...}{ignored}
}
\value{
character
}
\description{
Turn an \code{lzhd_hash} object into a base 64 encoded string
}
