% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min-hash-for-file.R
\name{as_lzjd_hash}
\alias{as_lzjd_hash}
\title{Turn an base 64 encoded LZJD hash into an \code{lzjd_hash} object}
\usage{
as_lzjd_hash(x)
}
\arguments{
\item{x}{base 64 encoded LZJD}
}
\value{
\code{lzjd_hash}
}
\description{
Turn an base 64 encoded LZJD hash into an \code{lzjd_hash} object
}
