% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min-hash-for-file.R
\name{compute_distance}
\alias{compute_distance}
\title{Compute the distance between two LZJD hashes}
\usage{
compute_distance(h1, h2)
}
\arguments{
\item{h1, h2}{LZJD hashes}
}
\value{
numeric distance value (0-1)
}
\description{
Compute the distance between two LZJD hashes
}
