% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min-hash-for-file.R
\name{compute_similarity}
\alias{compute_similarity}
\title{Compute the similarity between two LZJD hashes}
\usage{
compute_similarity(h1, h2)
}
\arguments{
\item{h1, h2}{LZJD hashes}
}
\value{
numeric similarity value (0-1)
}
\description{
Compute the similarity between two LZJD hashes
}
