% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lizzard-package.R
\docType{package}
\name{lizzard}
\alias{lizzard}
\alias{lizzard-package}
\title{Compute and Compare Lempel-Ziv Jaccard (LZJD) Similarity Hashes}
\description{
The Lempel-Ziv Jaccard Distance (LZJD) is a method to measure the similarity between binary byte sequences.
Tools are provided that wrap Java methods by Raff et al, (2017) \url{arXiv:1708.03346v2} for computing
and comparing LZJD hashes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/lizzard}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/lizzard/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
