% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min-hash-for-file.R
\name{min_hash_for_file}
\alias{min_hash_for_file}
\alias{print.lzjd_hash}
\title{Compute the LZJD hash for a file}
\usage{
min_hash_for_file(path)

\method{print}{lzjd_hash}(x, ...)
}
\arguments{
\item{path}{path to file (it will be \code{\link[=path.expand]{path.expand()}}ed).}

\item{x, }{an \code{lzhd_hash} object}

\item{...}{ignored}
}
\value{
a \code{lzjd_hash} object which corresponds to a \code{jsat.classifiers.DataPoint}
(\url{https://github.com/EdwardRaff/JSAT/blob/master/JSAT/src/jsat/classifiers/DataPoint.java)}
object which, in this case, is an integer vector.
}
\description{
Compute the LZJD hash for a file
}
\examples{
h1 <- min_hash_for_file(system.file("extdat", "index.html", package = "lizzard"))
}
