% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lizzard-package.R
\docType{package}
\name{lizzard}
\alias{lizzard}
\alias{lizzard-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/lizzard}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/lizzard/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
